Improve overall handling of icon images
Icon images loaded from the database are now evaluated for multi-DPI options and
the closest one to 128x128 is chosen, the others being discarded. Scaling is
performed if needed. Upon save all icons are converted to PNG.

Image manipulation methods are added to Tools.

Metadata::addCustomIcon now accepts raw image bytes to allow the selection
process to be performed. Metadata::addCustomIconScaled is removed as it was not
used anywhere.

Clarify some variable names, add missing includes and documentation
comments.